pyproject.toml
[project]
name = "central-log-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.69",
    "fastapi>=0.121.0",
    "fastmcp>=2.13.0.2",
    "kubernetes>=34.1.0",
    "langchain>=1.0.5",
    "langchain-core>=1.0.4",
    "langchain-groq>=1.0.0",
    "langgraph>=1.0.2",
    "motor>=3.7.1",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
